feat: create next endpoint for staffs get next appointment scheduled on their specialties, notifying users with real time hub

+++ src/MedScheduleProject/MedSchedule.Application/Responses/AppointmentResponse.cs
@@ -14,7 +14,7 @@ namespace MedSchedule.Application.Responses
         public Guid PatientId { get; set; }
         public ScheduleWorkResponse ScheduleWork { get; set; }
         public string SpecialtyName { get; set; }
-        public int Duration { get; set; }
+        public decimal Duration { get; set; }
         public EAppointmentStatus AppointmentStatus { get; set; }
     }
